Luckysheet-master-2.0.2/xuanxing/backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
pydantic==2.9.2
python-dotenv==1.0.1
gunicorn==23.0.0
dj-database-url==2.2.0
whitenoise==6.7.0
